Added admin check to edit & delete tool, edit & delete user. Added shed coord promote to User/api and User/models. Fixed syncdb script.
git-svn-id: svn+ssh://control.se.rit.edu/home/teams/261/s261-08d/svn/trunk@168 adf32048-93eb-11e3-923a-07678a7a3859

+++ penguin/mysite/apps/Tools/api_routes.py
@@ -18,41 +18,50 @@ def validate_fields(fields):
 
 @csrf_exempt
 def update(request):
+	#Update a tool
 	if request.method == "PUT":
 		put_data = json.loads(request.body.decode("utf-8"))
 		tool_id = int(put_data["id"])
-
-		if not validate_fields(put_data):
-			error = {"error": "One or more fields were left blank, make sure all fields are filled in before submitting."}
-			return HttpResponse(json.dumps(error), content_type="application/json", status=400)
-
-		# pickup arrangements field left blank -- revert to user's default pickup arrangements
 		current_user = User.get_user_by_username(request.session['user']['username'])
-		pickup_arrangement = current_user.default_pickup_arrangements
-		if put_data["tool_pickup_arrangements"]:
-			pickup_arrangement = put_data["tool_pickup_arrangements"]
+		
+		#Validate user is owner of tool, or admin
+		if ((current_user.is_admin==True) or (Tool.get_tool(tool_id).owner==current_user)):
+		
+			if not validate_fields(put_data):
+				error = {"error": "One or more fields were left blank, make sure all fields are filled in before submitting."}
+				return HttpResponse(json.dumps(error), content_type="application/json", status=400)
+	
+			# pickup arrangements field left blank -- revert to user's default pickup arrangements
+			
+			pickup_arrangement = current_user.default_pickup_arrangements
+			if put_data["tool_pickup_arrangements"]:
+				pickup_arrangement = put_data["tool_pickup_arrangements"]
+	
+			# convert str representing bool to actual bool
+			community_shed = False
+			if put_data["in_community_shed"] == "True":
+				community_shed = True
+	
+			# convert str representing bool to actual bool
+			tool_available = True
+			if put_data["tool_available"] == "False":
+				tool_available = False
+	
+			Tool.update_tool(tool_id,
+					put_data["name"],
+					put_data["description"],
+					put_data["tool_type"],
+					community_shed,
+					pickup_arrangement, 
+					tool_available)
+			tool = Tool.get_tool(tool_id)
+			return_tool = tool_to_json(tool)
+			return HttpResponse(json.dumps(return_tool), content_type="application/json")
+		else:
+			error = {"error": "You are not authorized to edit this tool."}
+			return HttpResponse(json.dumps(error), content_type="application/json", status=401)
 
-		# convert str representing bool to actual bool
-		community_shed = False
-		if put_data["in_community_shed"] == "True":
-			community_shed = True
-
-		# convert str representing bool to actual bool
-		tool_available = True
-		if put_data["tool_available"] == "False":
-			tool_available = False
-
-		Tool.update_tool(tool_id,
-				put_data["name"],
-				put_data["description"],
-				put_data["tool_type"],
-				community_shed,
-				pickup_arrangement, 
-				tool_available)
-		tool = Tool.get_tool(tool_id)
-		return_tool = tool_to_json(tool)
-		return HttpResponse(json.dumps(return_tool), content_type="application/json")
-
+	#Create a tool
 	if request.method == "POST":
 		post_data = json.loads(request.body.decode("utf-8"))
 		
@@ -106,13 +115,19 @@ def get_tool(request, tool_id):
 	  * Tool can be deleted while it's checked out.
 	"""
 	if request.method == "DELETE":
-		try:
-			Tool.delete_tool(tool_id)
-			returnmsg = { 'success': True }
-		except:
-			returnmsg = { 'success': False }
-			 
-		return HttpResponse(json.dumps(returnmsg), content_type="application/json")
+		#Validate user is owner of tool, or admin
+		current_user = User.get_user_by_username(request.session['user']['username'])
+		if ((current_user.is_admin==True) or (Tool.get_tool(tool_id).owner==current_user)):
+			try:
+				Tool.delete_tool(tool_id)
+				returnmsg = { 'success': True }
+			except:
+				returnmsg = { 'success': False }
+				 
+			return HttpResponse(json.dumps(returnmsg), content_type="application/json")
+		else:
+			error = {"error": "You are not authorized to edit this tool."}
+			return HttpResponse(json.dumps(error), content_type="application/json", status=401)
 
 @csrf_exempt
 def user_tools(request):
