+++ penguin/mysite/apps/Users/api_routes.py
@@ -99,29 +99,7 @@ def user(request):
 		return_user = user_to_json(new_user)
 		request.session['user'] = return_user
 		return HttpResponse(json.dumps(return_user), content_type="application/json")
-	"""
-		password mismatch
-		invalid zipcode
-		invalid email
-		invalid phone number
-		user is shed coorinator can't change zipcode
-		user has tools loaned out, can't change zip
-		user is borrowing tools, can't change zip
-	"""
-	if request.method == 'PUT':
-		put_data = json.loads(request.body.decode("utf-8"))
 
-		# invalid zipcode -- error 400
-		if (not validate_zip_code(put_data["zip_code"])):
-			error = {"error": "invalid zip code!"}
-			return HttpResponse(json.dumps(error), content_type="application/json", status=400)
-		
-		# invalid email -- error 400
-		try:
-			validate_email( put_data["email"] )
-		except ValidationError:
-			error = {"error": "You entered an invalid email address!"}
-			return HttpResponse(json.dumps(error), content_type="application/json", status=400)
 
 		# invalid phone number -- error 400
 		if not validate_phone_number(put_data["phone_number"]):
@@ -169,14 +147,87 @@ def userById(request, user_id):
 	  * Can delete users checking tools out.
 	"""
 	if request.method == "DELETE":
-		try:
-			User.delete_user(user_id)
-			returnmsg = { 'success': True }
-		except:
-			returnmsg = { 'success': False }
-			 
-		return HttpResponse(json.dumps(returnmsg), content_type="application/json")
+		current_user = request.session['user']['id']
+		if(user_id == current_user.id or current_user.is_admin==True):
+			try:
+				User.delete_user(user_id)
+				returnmsg = { 'success': True }
+			except:
+				returnmsg = { 'success': False }
+				 
+			return HttpResponse(json.dumps(returnmsg), content_type="application/json")
+		else:
+			error = {"error":"Unauthorized access."}
+			return HttpResponse(json.dumps(error), content_type="application/json", status=401)
+		
+		
+	"""
+	UPDATE USER
+	password mismatch
+	invalid zipcode
+	invalid email
+	invalid phone number
+	user is shed coorinator can't change zipcode
+	user has tools loaned out, can't change zip
+	user is borrowing tools, can't change zip
+	"""
+	if request.method == 'PUT':
+		put_data = json.loads(request.body.decode("utf-8"))
+		current_user = request.session['user']['id']
+		if(user_id == current_user.id or current_user.is_admin==True):
+			
+			# invalid zipcode -- error 400
+			if (not validate_zip_code(put_data["zip_code"])):
+				error = {"error": "invalid zip code!"}
+				return HttpResponse(json.dumps(error), content_type="application/json", status=400)
+			
+			# invalid email -- error 400
+			try:
+				validate_email( put_data["email"] )
+			except ValidationError:
+				error = {"error": "You entered an invalid email address!"}
+				return HttpResponse(json.dumps(error), content_type="application/json", status=400)
+	
+			# invalid phone number -- error 400
+			if not validate_phone_number(put_data["phone_number"]):
+				error = {"error": "You entered an invalid phone number!"}
+				return HttpResponse(json.dumps(error), content_type="application/json", status=400)
+	
+			if request.session['user']['zip_code'] != put_data['zip_code']:
+			# user is shed coordinator, can't change zip -- error 403
+				if request.session["user"]["is_shed_coordinator"]:
+					error = {"error": "You are currently shed coordinator, you may not change your zip code.  Please contact an admin."}
+					return HttpResponse(json.dumps(error), content_type="application/json", status=403)
+	
+			# user has tools being borrowed, can't change zip -- error 403
+				current_user = User.get_user_by_username(request.session['user']["username"])
+				if BorrowTransaction.get_borrow_transaction_user_owns(current_user.id):
+					error = {"error": "Some of your tools are currently being borrowed, you may not change your zip code.  Please contact an admin."}
+					return HttpResponse(json.dumps(error), content_type="application/json", status=403)
+	
+			# user is borrowing tools, can't change zip -- error 403
+				if BorrowTransaction.get_borrower_borrow_transactions(current_user.id):
+					error = {"error": "You are currently borrowing tools, you may not change your zip code.  Please contact an admin."}
+					return HttpResponse(json.dumps(error), content_type="application/json", status=403)
+	
+			# default pickup arrangements field left blank -- error 400
+			if not put_data["default_pickup_arrangements"]:
+				error ={"error": "pickup arrangements field was left blank.  Please specify your pickup arrangements"}
+				return HttpResponse(json.dumps(error), content_type="application/json", status=400)
+	
+			update_user = User.update_user( request.session['user']['username'],
+							put_data['phone_number'],
+							put_data['zip_code'],
+							put_data['email'],
+							put_data['default_pickup_arrangements'])
+			return_user = user_to_json(update_user)
+			request.session["user"] = return_user
+			return HttpResponse(json.dumps(return_user), content_type="application/json")
+		else:
+			error = {"error":"Unauthorized access."}
+			return HttpResponse(json.dumps(error), content_type="application/json", status=401)
 
+		
 @csrf_exempt
 def login(request):
 	"""
@@ -249,3 +300,23 @@ def get_admins(request):
 		for admin in admins:
 			return_list.append(user_to_json(admin))
 		return HttpResponse(json.dumps(return_list), content_type="application/json")
+
+@csrf_exempt
+def shedCoord(request):
+	if request.method == "PUT":
+		if (User.get_user(request.session['user']['id']).is_admin == True):
+			
+			put_data = json.loads(request.body.decode("utf-8"))
+			new_shed_coord = User.get_user(put_data['promote'])
+			zip_c = new_shed_coord.zip_code
+			
+			if (User.does_shed_coord_exist(zip_c)):
+				User.demote_user_from_shed_coordinator(User.get_shed_coordinator_for_zip(zip_c).id)
+		
+			User.promote_user_to_shed_coordinator(new_shed_coord.id)
+			
+			return_user = user_to_json(User.get_user(new_shed_coord.id))
+			return HttpResponse(json.dumps(return_user), content_type="application/json")
+		else:
+			error = {"error":"Unauthorized access."}
+			return HttpResponse(json.dumps(errror), content_type="application/json", status=401)
